@@ -3,16 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "lint": "^0.7.0",
-    "moment": "2.26.0",
-    "react": "^16.12.0",
+    "@reduxjs/toolkit": "^1.9.4",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "bootstrap": "^5.2.3",
+    "lint": "^0.8.19",
+    "moment": "2.29.4",
+    "react": "^18.2.0",
     "react-async": "^10.0.1",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.1"
+    "react-bootstrap": "^2.7.4",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "lint": "eslint src --ext .js",
@@ -38,6 +42,6 @@
     ]
   },
   "devDependencies": {
-    "npm-package-json-lint": "^5.1.0"
+    "npm-package-json-lint": "^6.4.0"
   }
 }
