@@ -1,3 +1,11 @@
-export type Ballpark = {
-    
-}
+export type TBallpark = {
+  id: string;
+  name: string;
+  city: string;
+  aka: string;
+  state: string;
+  league: string;
+  startDate: Date;
+  endDate: Date;
+  notes: string;
+};
